requirements.txt
Flask==0.7.2
Pillow
Numpy
https://storage.googleapis.com/tensorflow/linux/cpu/tensorflow-0.8.0-cp27-none-linux_x86_64.whl
